Remove direct redux devtools integration
You can now use the chrome extension, much simpler
