+++ package.json
@@ -69,9 +69,6 @@
     "pre-commit": "^1.1.2",
     "react-addons-test-utils": "^0.14.3",
     "react-router-bootstrap": "^0.20.1",
-    "redux-devtools": "^3.0.2",
-    "redux-devtools-dock-monitor": "^1.0.1",
-    "redux-devtools-log-monitor": "^1.0.2",
     "redux-logger": "^2.4.0",
     "stream-http": "^2.1.0",
     "style-loader": "^0.13.0",
